No more global setup for the connection since not all suites require a connection
